Address updates

+++ test/libsolidity/semanticTests/viaYul/dirty_calldata_struct.sol
@@ -9,7 +9,7 @@ contract C {
         assembly {
             r := x
         }
-        correct = r == 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff80;
+        correct = r == uint(0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff80);
     }
 }
 // ====
